+++ Changelog.md
@@ -3,6 +3,8 @@
 ### Added
 
 -   Digital channels can now be configured as triggers in the `Scope` pane.
+-   Option for cooshing a trigger bias level.
+-   Option for choosing the triggering edge.
 
 ## 4.2.1 - 2024-11-11